Extract terrain texture tile source reader
