+++ src/PlateauResoniteLink/Targets/Resonite/TerrainTextureSourceImage.cs
@@ -1,0 +1,18 @@
+using System;
+
+using PlateauResoniteLink.Domain.Importing;
+
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record TerrainTextureSourceImage(
+    Image<Rgba32> Image,
+    TextureUvRect? OccupiedUvRect) : IDisposable
+{
+    public void Dispose()
+    {
+        Image.Dispose();
+    }
+}
